feat: Implement useWebSearch composable for managing web search functionality and state in WebSearchNode component
